+++ src/library/DIPS.Mobile.UI/Components/Shell/Shell.cs
@@ -27,11 +27,15 @@ namespace DIPS.Mobile.UI.Components.Shell
             SetForegroundColor(this, Colors.GetColor(ForegroundColorName));
             SetTitleColor(this, Colors.GetColor(TitleTextColorName));
             
-            SetTabBarBackgroundColor(this, Colors.GetColor(ColorName.color_surface_subtle));
+            SetTabBarBackgroundColor(this, Colors.GetColor(ColorName.color_background_default));
             SetTabBarTitleColor(this, Colors.GetColor(ColorName.color_text_action));
             SetTabBarUnselectedColor(this, Colors.GetColor(ColorName.color_icon_subtle));
-            
+
+#if __ANDROID__
             SetNavBarHasShadow(this, false);
+#elif __IOS__
+            SetNavBarHasShadow(this, true);            
+#endif
         }
 
         private async void OnNavigated(object? sender, ShellNavigatedEventArgs e)